moved lights into own folder. added quad (area) light files.
