+++ programming_along/scene/Light/RotatingLight.cpp
