+++ programming_along/scene/Light/RotatingLight.h
